// source/frontend/src/models/api/generated/ApiGen_Concepts_FileChecklistItemType.ts
/**
 * File autogenerated by TsGenerator.
 * Do not manually modify, changes made to this file will be lost when this file is regenerated.
 */
import { ApiGen_Base_BaseConcurrent } from './ApiGen_Base_BaseConcurrent';

// LINK: @backend/apimodels/Models/Concepts/File/FileChecklistItemTypeModel.cs
export interface ApiGen_Concepts_FileChecklistItemType extends ApiGen_Base_BaseConcurrent {
  code: string | null;
  sectionCode: string | null;
  description: string | null;
  hint: string | null;
  isDisabled: boolean;
  displayOrder: number | null;
}